Public API

Functions

NormalHermiteSplines.interpolate — Function

interpolate(nodes::Matrix{T}, values::Vector{T}, kernel::RK = RK_H0()) where {T <: AbstractFloat, RK <: ReproducingKernel_0}

Create a normal spline by values of function defined at nodes.

Arguments

nodes: Locations where function values are defined. This should be an n×n_1 matrix, where n is dimension of the sampled space and n_1 is the number of function value nodes. It means that each column in the matrix defines one node.

values::Vector{T}: function values at n_1 interpolation nodes.

kernel: reproducing kernel of Bessel potential space the normal spline is constructed in. It must be a struct object of the following type: RK_H0 if the spline is constructing as a continuous function, RK_H1 if the spline is constructing as a differentiable function, RK_H2 if the spline is constructing as a twice differentiable function.

Return: a completely initialized NormalSpline object that can be passed to evaluate function to interpolate the data to required points.

interpolate(nodes::Matrix{T}, values::Vector{T}, d_nodes::Matrix{T}, es::Matrix{T}, d_values::Vector{T}, kernel::RK = RK_H0()) where {T <: AbstractFloat, RK <: ReproducingKernel_0}

Create a normal spline by values of function defined at nodes and d_values of function directional derivatives defined at d_nodes.

Arguments

nodes: Locations where function values are defined. This should be an n×n_1 matrix, where n is dimension of the sampled space and n_1 is the number of function value nodes. It means that each column in the matrix defines one node.

values::Vector{T}: function values at n_1 interpolation nodes.

d_nodes: Locations where function directional derivatives are defined. This should be an n×n_2 matrix, where n is dimension of the sampled space and n_2 is the number of function directional derivative nodes.

es: Directions of the function directional derivatives. This should be an n×n_2 matrix, where n is dimension of the sampled space and n_2 is the number of function directional derivative nodes. It means that each column in the matrix defines one direction of the function directional derivative.

d_values::Vector{T}: function directional derivative values at n_2 interpolation nodes.

kernel: reproducing kernel of Bessel potential space the normal spline is constructed in. It must be a struct object of the following type: RK_H1 if the spline is constructing as a differentiable function, RK_H2 if the spline is constructing as a twice differentiable function.

Return: a completely initialized NormalSpline object that can be passed to evaluate function to interpolate the data to required points.

NormalHermiteSplines.prepare — Function

prepare(nodes::Matrix{T}, kernel::RK = RK_H0()) where {T <: AbstractFloat, RK <: ReproducingKernel_0}

Prepare a normal spline by constructing and factorizing Gram matrix of the interpolation problem. Initialize the NormalSpline object.

Arguments

nodes: Locations where function values are defined. This should be an n×n_1 matrix, where n is dimension of the sampled space and n_1 is the number of function value nodes. It means that each column in the matrix defines one node.

kernel: reproducing kernel of Bessel potential space the normal spline is constructed in. It must be a struct object of the following type: RK_H0 if the spline is constructing as a continuous function, RK_H1 if the spline is constructing as a differentiable function, RK_H2 if the spline is constructing as a twice differentiable function.

Return: a partly initialized NormalSpline object that must be passed to construct function in order to complete the spline initialization.

prepare(nodes::Matrix{T}, d_nodes::Matrix{T}, es::Matrix{T}, kernel::RK = RK_H1()) where {T <: AbstractFloat, RK <: ReproducingKernel_1}

Prepare a normal spline by constructing and factorizing Gram matrix of the interpolation problem. Initialize the NormalSpline object.

Arguments

nodes: Locations where function values are defined. This should be an n×n_1 matrix, where n is dimension of the sampled space and n_1 is the number of function value nodes. It means that each column in the matrix defines one node.

d_nodes: Locations where function directional derivatives are defined. This should be an n×n_2 matrix, where n is dimension of the sampled space and n_2 is the number of function directional derivative nodes.

es: Directions of the function directional derivatives. This should be an n×n_2 matrix, where n is dimension of the sampled space and n_2 is the number of function directional derivative nodes. It means that each column in the matrix defines one direction of the function directional derivative.

kernel: reproducing kernel of Bessel potential space the normal spline is constructed in. It must be a struct object of the following type: RK_H1 if the spline is constructing as a differentiable function, RK_H2 if the spline is constructing as a twice differentiable function.

Return: a partly initialized NormalSpline object that must be passed to construct function in order to complete the spline initialization.

NormalHermiteSplines.construct — Function

construct(spline::NormalSpline{T, RK}, values::Vector{T}) where {T <: AbstractFloat, RK <: ReproducingKernel_0}

Construct an interpolating spline by calculating coefficients of the spline and completely initializing the NormalSpline object.

Arguments

spline::NormalSpline{T, RK}: a partly initialized NormalSpline object returned by prepare function.

values::Vector{T}: function values at n_1 interpolation nodes.

Return: a completely initialized NormalSpline object that can be passed to evaluate function to interpolate the data to required points.

construct(spline::NormalSpline{T, RK}, values::Vector{T}, d_values::Vector{T}) where {T <: AbstractFloat, RK <: ReproducingKernel_0}

Construct an interpolating spline by calculating coefficients of the spline and completely initializing the NormalSpline object.

Arguments

spline::NormalSpline{T, RK}: a partly initialized NormalSpline object returned by prepare function.

values::Vector{T}: function values at n_1 interpolation nodes.

d_values::Vector{T}: function directional derivative values at n_2 interpolation nodes.

Return: a completely initialized NormalSpline object that can be passed to evaluate function to interpolate the data to required points.

NormalHermiteSplines.evaluate — Function

evaluate(spline::NormalSpline{T, RK}, points::Matrix{T}) where {T <: AbstractFloat, RK <: ReproducingKernel_0}

Evaluate normal spline at the locations defined in points.

Arguments

spline::NormalSpline{T, RK}: a NormalSpline object returned by interpolate or construct function.

points::Matrix{T}: locations at which spline values are evaluating. This should be an n×m matrix, where n is dimension of the sampled space and m is the number of locations where spline values are evaluating. It means that each column in the matrix defines one location.

Return: Vector{T} of spline values at the locations defined in points.

evaluate(spline::NormalSpline{T, RK}, point::Vector{T}) where {T <: AbstractFloat, RK <: ReproducingKernel_0}

Evaluate normal spline at the location defined in point.

Arguments

spline::NormalSpline{T, RK}: a NormalSpline object returned by interpolate or construct function.

point::Vector{T}: location at which spline value is evaluating. This should be a vector of size n, where n is dimension of the sampled space.

Return: spline value at the location defined in point.

NormalHermiteSplines.evaluate_grad — Function

evaluate_grad(spline::NormalSpline{T, RK}, point::Vector{T}) where {T <: AbstractFloat, RK <: ReproducingKernel_0}

Evaluate a gradient of the normal spline at the location defined in point.

Arguments

spline::NormalSpline{T, RK}: a NormalSpline object returned by interpolate or construct function.

point::Vector{T}: location at which gradient value is evaluating. This should be a vector of size n, where n is dimension of the sampled space.

Return: Vector{T} - gradient of the normal spline at the location defined in point.

NormalHermiteSplines.estimate_interpolation_quality — Function

estimate_interpolation_quality(spline::NormalSpline{T, RK}) where {T <: AbstractFloat, RK <: ReproducingKernel_0}

Estimate the interpolation quality

Arguments

spline::NormalSpline{T, RK}: a NormalSpline object returned by construct or interpolate function.

Return: RMSE of interpolation at function value nodes.

NormalHermiteSplines.get_cond — Function

get_cond(spline::NormalSpline{T, RK}) where {T <: AbstractFloat, RK <: ReproducingKernel_0}

Get an estimation of the Gram matrix condition number. (C. Brás, W. Hager, J. Júdice, An investigation of feasible descent algorithms for estimating the condition number of a matrix. TOP Vol.20, No.3, 2012.)

Arguments

spline::NormalSpline{T, RK}: a NormalSpline object returned by prepare, construct or interpolate function.

Return: an estimation of the Gram matrix condition number.

NormalHermiteSplines.get_epsilon — Function

get_epsilon(spline::NormalSpline{T, RK}) where {T <: AbstractFloat, RK <: ReproducingKernel_0}

Get a ε - 'scaling parameter' of Bessel Potential space the normal spline was built in.

Arguments

spline::NormalSpline{T, RK}: a NormalSpline object returned by prepare, construct or interpolate function.

Return: ε.

NormalHermiteSplines.estimate_epsilon — Function

estimate_epsilon(nodes::Matrix{T}) where T <: AbstractFloat

Get an the estimation of the 'scaling parameter' of Bessel Potential space the normal spline was built in. This coinsides with result returned by get_epsilon function if all nodes are located in a unit hypercube, othewise the estimated value of ε can be significantly higher then necessary.

Arguments

nodes: Locations where function values are defined. This should be an n×n_1 matrix, where n is dimension of the sampled space and n_1 is the number of function value nodes. It means that each column in the matrix defines one node.

Return: estimation of ε.

estimate_epsilon(nodes::Matrix{T}, d_nodes::Matrix{T}) where T <: AbstractFloat

Get an the estimation of the 'scaling parameter' of Bessel Potential space the normal spline was built in. This coinsides with result returned by get_epsilon function if all nodes and d_nodes are located in a unit hypercube, othewise the estimated value of ε can be significantly higher then necessary.

Arguments

nodes: Locations where function values are defined. This should be an n×n_1 matrix, where n is dimension of the sampled space and n_1 is the number of function value nodes. It means that each column in the matrix defines one node.

d_nodes: Locations where function directional derivatives are defined. This should be an n×n_2 matrix, where n is dimension of the sampled space and n_2 is the number of function directional derivative nodes.

Return: estimation of ε.

Types

Bessel Potential space Reproducing Kernels

NormalHermiteSplines.RK_H0 — Type

struct RK_H0{T <: AbstractFloat} <: ReproducingKernel_0

Define a type of reproducing kernel of Bessel Potential space $H^{n/2 + 1/2}_ε (R^n)$:

\[V(\eta , \xi, \varepsilon) = \exp (-\varepsilon |\xi - \eta|) \, .\]

Fields

ε::T: 'scaling parameter' from the Bessel Potential space definition, it may be omitted in the struct constructor otherwise it must be greater than zero

NormalHermiteSplines.RK_H1 — Type

struct RK_H1{T <: AbstractFloat} <: ReproducingKernel_1

Define a type of reproducing kernel of Bessel Potential space $H^{n/2 + 3/2}_ε (R^n)$:

\[V(\eta , \xi, \varepsilon) = \exp (-\varepsilon |\xi - \eta|) (1 + \varepsilon |\xi - \eta|) \, .\]

Fields

ε::T: 'scaling parameter' from the Bessel Potential space definition, it may be omitted in the struct constructor otherwise it must be greater than zero

NormalHermiteSplines.RK_H2 — Type

struct RK_H2{T <: AbstractFloat} <: ReproducingKernel_2

Define a type of reproducing kernel of Bessel Potential space $H^{n/2 + 5/2}_ε (R^n)$:

\[V(\eta , \xi, \varepsilon) = \exp (-\varepsilon |\xi - \eta|) (3 + 3\varepsilon |\xi - \eta| + \varepsilon ^2 |\xi - \eta| ^2 ) \, .\]

Fields

ε::T: 'scaling parameter' from the Bessel Potential space definition, it may be omitted in the struct constructor otherwise it must be greater than zero

NormalSpline structure

NormalHermiteSplines.NormalSpline — Type

struct NormalSpline{T, RK} <: AbstractSpline where {T <: AbstractFloat, RK <: ReproducingKernel_0}

Define a structure containing full information of a normal spline

Fields

_kernel: a reproducing kernel spline was built with.

_compression: factor of transforming the original node locations into unit hypercube

_nodes: transformed function value nodes

_values: function values at interpolation nodes

_d_nodes: transformed function derivative nodes

_es: normalized derivative directions

_d_values: function directional derivative values

_min_bound: minimal bounds of the original node locations area

_gram: Gram matrix of the problem

_chol: Cholesky factorization of the Gram matrix

_mu: spline coefficients

_cond: estimation of the Gram matrix condition number

Index

NormalHermiteSplines.construct

NormalHermiteSplines.estimate_epsilon

NormalHermiteSplines.estimate_interpolation_quality

NormalHermiteSplines.evaluate

NormalHermiteSplines.evaluate_grad

NormalHermiteSplines.get_cond

NormalHermiteSplines.get_epsilon

NormalHermiteSplines.interpolate

NormalHermiteSplines.prepare

NormalHermiteSplines.NormalSpline

NormalHermiteSplines.RK_H0

NormalHermiteSplines.RK_H1

NormalHermiteSplines.RK_H2